Abstracted styles from List, ListItem and ListText. Fixed some import errors in other components
