+++ packages/react-atlas-core/src/utils/index.js
@@ -2,5 +2,6 @@ import events from './events';
 import prefixer from './prefixer';
 import time from './time';
 import utils from './utils';
+import classNames from './classNames';
 
-export {events, prefixer, time, utils};
+export {events, prefixer, time, utils, classNames};
