+++ packages/react-atlas-default-theme/src/index.js
@@ -13,3 +13,6 @@ export Dialog from './dialog';
 export Drawer from './drawer';
 export Header from './header';
 export Input from './input';
+export List from './list';
+export ListItem from './list';
+export ListText from './list';
