+++ packages/react-atlas-default-theme/src/list/List.js
@@ -1,0 +1,5 @@
+import { List } from 'react-atlas-core';
+import style from './list.css';
+import createComponent from '../createComponent';
+
+export default createComponent(List, style);
