+++ packages/react-atlas-default-theme/src/list/ListItem.js
@@ -1,0 +1,5 @@
+import { ListItem } from 'react-atlas-core';
+import style from './list.css';
+import createComponent from '../createComponent';
+
+export default createComponent(ListItem, style);
