+++ packages/react-atlas-default-theme/src/list/ListText.js
@@ -1,0 +1,5 @@
+import { ListText } from 'react-atlas-core';
+import style from './list.css';
+import createComponent from '../createComponent';
+
+export default createComponent(ListText, style);
